restaurantapp/requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.10
python-decouple==3.8
pillow==11.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0